secmet: split qualifiers.py into a qualifiers module
