+++ antismash/common/secmet/qualifiers/__init__.py
@@ -1,0 +1,11 @@
+# License: GNU Affero General Public License v3 or later
+# A copy of GNU AGPL v3 should have been included in this software package in LICENSE.txt.
+
+""" Classes representing complex qualifiers for features.
+"""
+
+from .asf import ActiveSiteFinderQualifier
+from .gene_functions import GeneFunction, GeneFunctionAnnotations
+from .go import GOQualifier
+from .nrps_pks import NRPSPKSQualifier
+from .secmet import SecMetQualifier
